USHIFT-936: split up large config.go
Move data structures and related functions into their own files.
