+++ pkg/config/dns.go
@@ -1,0 +1,14 @@
+package config
+
+type DNS struct {
+	// baseDomain is the base domain of the cluster. All managed DNS records will
+	// be sub-domains of this base.
+	//
+	// For example, given the base domain `example.com`, router exposed
+	// domains will be formed as `*.apps.example.com` by default,
+	// and API service will have a DNS entry for `api.example.com`,
+	// as well as "api-int.example.com" for internal k8s API access.
+	//
+	// Once set, this field cannot be changed.
+	BaseDomain string `json:"baseDomain"`
+}
