+++ pkg/config/flags.go
@@ -1,0 +1,23 @@
+package config
+
+import (
+	"github.com/spf13/pflag"
+	"k8s.io/apimachinery/pkg/util/sets"
+	"k8s.io/component-base/logs"
+)
+
+func HideUnsupportedFlags(flags *pflag.FlagSet) {
+	// hide logging flags that we do not use/support
+	loggingFlags := pflag.NewFlagSet("logging-flags", pflag.ContinueOnError)
+	logs.AddFlags(loggingFlags)
+
+	supportedLoggingFlags := sets.NewString("v")
+
+	loggingFlags.VisitAll(func(pf *pflag.Flag) {
+		if !supportedLoggingFlags.Has(pf.Name) {
+			flags.MarkHidden(pf.Name)
+		}
+	})
+
+	flags.MarkHidden("version")
+}
