+++ pkg/config/ingress.go
@@ -1,0 +1,6 @@
+package config
+
+type IngressConfig struct {
+	ServingCertificate []byte
+	ServingKey         []byte
+}
