+++ pkg/config/network.go
@@ -1,0 +1,63 @@
+package config
+
+import (
+	"fmt"
+	"net"
+
+	"github.com/apparentlymart/go-cidr/cidr"
+)
+
+type Network struct {
+	// IP address pool to use for pod IPs.
+	// This field is immutable after installation.
+	ClusterNetwork []ClusterNetworkEntry `json:"clusterNetwork"`
+
+	// IP address pool for services.
+	// Currently, we only support a single entry here.
+	// This field is immutable after installation.
+	ServiceNetwork []string `json:"serviceNetwork"`
+
+	// The port range allowed for Services of type NodePort.
+	// If not specified, the default of 30000-32767 will be used.
+	// Such Services without a NodePort specified will have one
+	// automatically allocated from this range.
+	// This parameter can be updated after the cluster is
+	// installed.
+	// +kubebuilder:validation:Pattern=`^([0-9]{1,4}|[1-5][0-9]{4}|6[0-4][0-9]{3}|65[0-4][0-9]{2}|655[0-2][0-9]|6553[0-5])-([0-9]{1,4}|[1-5][0-9]{4}|6[0-4][0-9]{3}|65[0-4][0-9]{2}|655[0-2][0-9]|6553[0-5])$`
+	ServiceNodePortRange string `json:"serviceNodePortRange"`
+
+	// The DNS server to use
+	DNS string `json:"-"`
+}
+
+type ClusterNetworkEntry struct {
+	// The complete block for pod IPs.
+	CIDR string `json:"cidr"`
+}
+
+func (c *Config) computeAndUpdateClusterDNS() error {
+	if len(c.Network.ServiceNetwork) == 0 {
+		return fmt.Errorf("network.serviceNetwork not filled in")
+	}
+
+	clusterDNS, err := getClusterDNS(c.Network.ServiceNetwork[0])
+	if err != nil {
+		return fmt.Errorf("failed to get DNS IP: %v", err)
+	}
+	c.Network.DNS = clusterDNS
+	return nil
+}
+
+// getClusterDNS returns cluster DNS IP that is 10th IP of the ServiceNetwork
+func getClusterDNS(serviceCIDR string) (string, error) {
+	_, service, err := net.ParseCIDR(serviceCIDR)
+	if err != nil {
+		return "", fmt.Errorf("invalid service cidr %v: %v", serviceCIDR, err)
+	}
+	dnsClusterIP, err := cidr.Host(service, 10)
+	if err != nil {
+		return "", fmt.Errorf("service cidr must have at least 10 distinct host addresses %v: %v", serviceCIDR, err)
+	}
+
+	return dnsClusterIP.String(), nil
+}
