+++ pkg/config/util.go
@@ -1,0 +1,21 @@
+package config
+
+func StringInList(s string, list []string) bool {
+	for _, x := range list {
+		if x == s {
+			return true
+		}
+	}
+	return false
+}
+
+func stringSliceContains(list []string, elements ...string) bool {
+	for _, value := range list {
+		for _, element := range elements {
+			if value == element {
+				return true
+			}
+		}
+	}
+	return false
+}
